# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RankedStudentResponse": {
        "description": "A student in a ranking.",
        "properties": {
          "average_percentage": {
            "title": "Average Percentage",
            "type": "number"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "student_id": {
            "title": "Student Id",
            "type": "string"
          }
        },
        "required": [
          "student_id",
          "name",
          "average_percentage"
        ],
        "title": "RankedStudentResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "\nStudent grade management: students, courses, enrolments, grades, reports and an\nAI assistant.\n\n### Errors\n\nEvery failure returns `application/problem+json` (RFC 9457) carrying a stable\nmachine-readable `code` and a structured `context`:\n\n```json\n{ \"type\": \"about:blank#STUDENT_NOT_FOUND\", \"status\": 404,\n  \"code\": \"STUDENT_NOT_FOUND\", \"detail\": \"No student with id 'S999'.\",\n  \"context\": { \"student_id\": \"S999\" } }\n```\n\n**Clients should render the `code`, not the `detail`.** `detail` is a\ndeveloper-facing string and is only ever English; the interface translates the code\ninto the reader's language. This is why the API ships no message catalogue.\n\n### Authorization\n\nAuthentication is a session cookie (`HttpOnly`, `SameSite=Lax`). Access is filtered\nin the query rather than checked in the handler, so what you can see depends on your\nrole:\n\n| Role | Students | Courses | Grades |\n|---|---|---|---|\n| `student` | only themselves | those they are enrolled on | only their own |\n| `teacher` | those enrolled on a course they own | those they own | within their own courses |\n| `admin` | all | all | all |\n| `superadmin` | all, plus AI provider configuration | all | all |\n\nA list endpoint returns what you may see, never a `403`. Only *actions* beyond your\nrole return `403`.\n",
    "title": "Grade Tracker API",
    "version": "2.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/analytics/at-risk": {
      "get": {
        "description": "Worst first, since that is the order an intervention list is read in. Ungraded students are excluded: no data is not the same as poor performance.",
        "operationId": "at_risk_analytics_at_risk_get",
        "parameters": [
          {
            "in": "query",
            "name": "threshold",
            "required": false,
            "schema": {
              "default": 60.0,
              "maximum": 100,
              "minimum": 0,
              "title": "Threshold",
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RankedStudentResponse"
                  },
                  "title": "Response At Risk Analytics At Risk Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Students below a threshold",
        "tags": [
          "Analytics"
        ]
      }
    },
...
    "/analytics/top-students": {
      "get": {
        "description": "Students with no grades are excluded rather than ranked as zero.",
        "operationId": "top_students_analytics_top_students_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 5,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RankedStudentResponse"
                  },
                  "title": "Response Top Students Analytics Top Students Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Highest-averaging students",
        "tags": [
          "Analytics"
        ]
      }
    },
...
    "/reports/course/{course_id}": {
      "get": {
        "description": "A student requesting this sees the class statistics but only their own marks.",
        "operationId": "course_report_reports_course__course_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "course_id",
            "required": true,
            "schema": {
              "title": "Course Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Course Report Reports Course  Course Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "`COURSE_NOT_FOUND`."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "A course's report",
        "tags": [
          "Reports"
        ]
      }
    },
    "/reports/student/{student_id}": {
      "get": {
        "description": "Numbers and identifiers only — no sentences. The client renders the wording.\n\nA student's own report is complete. A teacher's copy contains only the grades from courses they own, so the report cannot expose marks a scoped list endpoint would have hidden.",
        "operationId": "student_report_reports_student__student_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "student_id",
            "required": true,
            "schema": {
              "title": "Student Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Student Report Reports Student  Student Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "`STUDENT_NOT_FOUND` — unknown, or outside your scope."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "A student's report",
        "tags": [
          "Reports"
        ]
      }
    },
    "/reports/summary": {
      "get": {
        "description": "Totals, distribution, top students and at-risk students. Staff only: a summary over every student cannot be meaningfully scoped to one of them.",
        "operationId": "summary_report_reports_summary_get",
        "parameters": [
          {
            "description": "Percentage below which a student is at risk.",
            "in": "query",
            "name": "at_risk_threshold",
            "required": false,
            "schema": {
              "default": 60.0,
              "description": "Percentage below which a student is at risk.",
              "maximum": 100,
              "minimum": 0,
              "title": "At Risk Threshold",
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Summary Report Reports Summary Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Institution-wide summary",
        "tags": [
          "Reports"
        ]
      }
    },
    "/reports/{kind}/{entity_id}/export.csv": {
      "get": {
        "description": "The one endpoint that returns translated prose, because a downloaded file has no frontend to render it. `?locale=de` or `fr` also switches the delimiter to `;`, which is what Excel expects in those locales — a comma-separated file opens there as a single column.\n\nUse `summary` as both `kind` and `entity_id` for the institution-wide report.",
        "operationId": "export_csv_reports__kind___entity_id__export_csv_get",
        "parameters": [
          {
            "in": "path",
            "name": "kind",
            "required": true,
            "schema": {
              "title": "Kind",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "entity_id",
            "required": true,
            "schema": {
              "title": "Entity Id",
              "type": "string"
            }
          },
          {
            "description": "Language for CSV column headers.",
            "in": "query",
            "name": "locale",
            "required": false,
            "schema": {
              "default": "en",
              "description": "Language for CSV column headers.",
              "examples": [
                "de"
              ],
              "title": "Locale",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/csv": {},
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "The CSV file."
          },
          "404": {
            "description": "`STUDENT_NOT_FOUND` or `COURSE_NOT_FOUND`."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Download a report as CSV",
        "tags": [
          "Reports"
        ]
      }
    },
...
  },
...
}